iii
git-svn-id: svn://192.168.0.99/delfols@5026 5ada3c48-002d-0410-850d-99865cc6fd0f

+++ delfols/FormMain.cpp
@@ -54,6 +54,12 @@ namespace delfols {
 
 		for each(DirectoryInfo^ subdir in subdirs)
 		{
+			if ((subdir->Attributes & System::IO::FileAttributes::ReparsePoint)==System::IO::FileAttributes::ReparsePoint)
+			{
+				theDeleteDir(subdir->FullName);
+				continue;
+			}
+			
 			deleteAll(subdir);
 			theDeleteDir(subdir->FullName);
 		}
